Change USART2 to USART1, motor controls now independent, add Ignore CommandMode, add SVD
USART2 to USART1 to match change in PCB layout

Originally motor of id 0x143 depends on the response of motor 0x142, now changed to be done by a timer delay instead. This is to ensure that we can control just 0x143 without depending on 0x142 working.

Add Ignore CommandMode to ignore any motor such that no CAN message is sent for that motor.